requirements.txt
fastapi
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
uvicorn
python-multipart
email-validator

